refactor: flatted comments provide context for consume in comment rendering layer
