+++ packages/script/zenza-watch/src/contexts/comment-context.ts
@@ -1,0 +1,6 @@
+import {createContext} from "@lit/context";
+import type {FlattedComment} from "@/comment-list";
+
+export const commentContext = createContext<FlattedComment[] | undefined>(
+  "comment",
+);
